refactor(app): split settings screen panels
